+++ test/test_process_config.py
@@ -13,8 +13,8 @@ MAIN_CONFIG_KEYS = ["buttonSize","trialsPerItem","conditionId","relativePosition
 @pytest.mark.parametrize(
     "f, expected",[
         (Path(__file__).parent / "test_files/working_file.toml", True),
-        (Path(__file__).parent / "test_files/working_file.toml", True),
         (Path(__file__).parent / "test_files/working_file_6.toml", True),
+        (Path(__file__).parent / "test_files/working_file_7.toml", True),
         (Path(__file__).parent / "test_files/working_file.expconfig", True),
         (Path(__file__).parent / "test_files/working_file_2.expconfig", True),
         (Path(__file__).parent / "test_files/working_file_3.expconfig", True),
@@ -46,7 +46,6 @@ def _test_func(config):
 
 @pytest.mark.parametrize(
     "f, expected",[
-        (Path(__file__).parent / "test_files/working_file.toml", True),
         (Path(__file__).parent / "test_files/working_file.toml", True),
         (Path(__file__).parent / "test_files/working_file.expconfig", True),
         (Path(__file__).parent / "test_files/working_file_2.expconfig", False),